+++ xknx/io/request_response/connect.py
@@ -6,10 +6,11 @@ from typing import TYPE_CHECKING
 from xknx.knxip import (
     HPAI,
     ConnectRequest,
-    ConnectRequestType,
+    ConnectRequestInformation,
     ConnectResponse,
     KNXIPFrame,
 )
+from xknx.telegram import IndividualAddress
 
 from .request_response import RequestResponse
 
@@ -18,27 +19,34 @@ if TYPE_CHECKING:
 
 
 class Connect(RequestResponse):
-    """Class to send a ConnectRequest and wait for ConnectResponse."""
+    """
+    Class to send a ConnectRequest and wait for ConnectResponse.
+
+    Setting a `individual_address` is only supported for Tunnelling v2 connections.
+    """
 
     def __init__(
         self,
         transport: KNXIPTransport,
         local_hpai: HPAI,
+        individual_address: IndividualAddress | None = None,
     ):
         """Initialize Connect class."""
         super().__init__(transport, ConnectResponse)
         self.communication_channel = 0
         self.data_endpoint = HPAI()
         self.identifier = 0
+        self.individual_address = individual_address
         self.local_hpai = local_hpai
 
     def create_knxipframe(self) -> KNXIPFrame:
         """Create KNX/IP Frame object to be sent to device."""
+        cri = ConnectRequestInformation(individual_address=self.individual_address)
         # use the same HPAI for control_endpoint and data_endpoint
         connect_request = ConnectRequest(
-            request_type=ConnectRequestType.TUNNEL_CONNECTION,
             control_endpoint=self.local_hpai,
             data_endpoint=self.local_hpai,
+            cri=cri,
         )
         return KNXIPFrame.init_from_body(connect_request)